# docs/swagger.yaml
basePath: /
definitions:
  main.Activity:
    description: 配對活動資訊
    properties:
      created_by:
        type: integer
      description:
        type: string
      id:
        type: integer
      location:
        $ref: '#/definitions/main.Location'
      location_id:
        type: integer
      target_count:
        type: integer
      title:
        type: string
    type: object
...
  main.Location:
    description: 地點資訊
    properties:
      address:
        type: string
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
    type: object
  main.Match:
    description: 配對局資訊
    properties:
      activity:
        $ref: '#/definitions/main.Activity'
      activity_id:
        type: integer
      id:
        type: integer
      match_time:
        type: string
      organizer:
        $ref: '#/definitions/main.User'
      organizer_id:
        type: integer
      status:
        description: open, closed, completed
        type: string
    type: object
  main.MatchParticipant:
    description: 配對參與者資訊
    properties:
      id:
        type: integer
      joined_at:
        type: string
      match:
        $ref: '#/definitions/main.Match'
      match_id:
        type: integer
      status:
        description: pending, approved, rejected
        type: string
      user:
        $ref: '#/definitions/main.User'
      user_id:
        type: integer
    type: object
  main.Review:
    description: 評分與留言資訊
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      match:
        $ref: '#/definitions/main.Match'
      match_id:
        type: integer
      reviewee:
        $ref: '#/definitions/main.User'
      reviewee_id:
        type: integer
      reviewer:
        $ref: '#/definitions/main.User'
      reviewer_id:
        type: integer
      score:
        description: 3-5分
        type: integer
    type: object
  main.ReviewLike:
    description: 評論點讚/倒讚資訊
    properties:
      id:
        type: integer
      is_like:
        description: 'true: 點讚, false: 倒讚'
        type: boolean
      review:
        $ref: '#/definitions/main.Review'
      review_id:
        type: integer
      user:
        $ref: '#/definitions/main.User'
      user_id:
        type: integer
    type: object
  main.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: integer
      email:
        type: string
      id:
        type: integer
      is_admin:
        description: 添加管理員標記
        type: boolean
      name:
        type: string
      social_id:
        type: string
      social_provider:
        type: string
      updated_at:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 這是一個買一送一配對網站的API文檔
  title: 買一送一配對網站 API
  version: "1.0"
paths:
...
  /organizer/matches/{id}/participants/{participant_id}/approve:
    put:
      consumes:
      - application/json
      description: 開局者審核通過指定配對局的參與者
      parameters:
      - description: 配對局ID
        in: path
        name: id
        required: true
        type: integer
      - description: 參與者ID
        in: path
        name: participant_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.MatchParticipant'
        "400":
          description: 無效的配對局 ID 或參與者 ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 無法審核通過參與者
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 審核通過參與者
      tags:
      - 開局者
  /organizer/matches/{id}/participants/{participant_id}/reject:
    put:
      consumes:
      - application/json
      description: 開局者審核拒絕指定配對局的參與者
      parameters:
      - description: 配對局ID
        in: path
        name: id
        required: true
        type: integer
      - description: 參與者ID
        in: path
        name: participant_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.MatchParticipant'
        "400":
          description: 無效的配對局 ID 或參與者 ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 無法審核拒絕參與者
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 審核拒絕參與者
      tags:
      - 開局者
...
  /review-like/reviews/{id}/dislike:
    post:
      consumes:
      - application/json
      description: 對指定ID的評論進行倒讚
      parameters:
      - description: 評論ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.ReviewLike'
        "400":
          description: 無效的評論 ID 或已倒讚
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 無法倒讚評論
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 倒讚評論
      tags:
      - 評論
  /review-like/reviews/{id}/like:
    post:
      consumes:
      - application/json
      description: 對指定ID的評論進行點讚
      parameters:
      - description: 評論ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.ReviewLike'
        "400":
          description: 無效的評論 ID 或已點讚
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 無法點讚評論
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 點讚評論
      tags:
      - 評論
  /review/matches/{id}:
    post:
      consumes:
      - application/json
      description: 建立對指定配對局中其他參與者的評分與留言
      parameters:
      - description: 配對局ID
        in: path
        name: id
        required: true
        type: integer
      - description: 評分與留言資訊
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/main.Review'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Review'
        "400":
          description: 無效的請求資料或已評分過
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 無法建立評分記錄
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 建立評分與留言
      tags:
      - 評分
...
securityDefinitions:
  ApiKeyAuth:
    description: 輸入 'Bearer &lt;JWT token&gt;' 進行認證。可以先透過 Facebook 登入取得 token。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"